authentication-spring-security: add exception handler, refactor returned values of methods of create on each entity

+++ dao/src/main/java/com/epam/esm/hibernate/CertificateRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface CertificateRepository extends Repository<GiftCertificateEntity> {
 
     @Transactional
-    int create(GiftCertificateEntity entity);
+    long create(GiftCertificateEntity entity);
 
     @Transactional
     void delete(long id);
