+++ dao/src/main/java/com/epam/esm/hibernate/TagRepository.java
@@ -7,7 +7,7 @@ import org.springframework.transaction.annotation.Transactional;
 public interface TagRepository extends Repository<TagEntity> {
 
     @Transactional
-    int create(TagEntity entity);
+    long create(TagEntity entity);
 
     @Transactional
     void delete(long id);
